requirements.txt
pytest
selene>=2.0.0b16
allure-pytest==2.12.0